Organization App Team Model  views.py file  Data Security work and so…

+++ organization/views.py
@@ -1,13 +1,19 @@
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Team
 from .forms import TeamForm
+import logging
+from .decorators import branch_owner_required
+logger = logging.getLogger(__name__)
 
 
 def home(request):
     return render(request, 'index.html')
 
+#Create Team
 @login_required
+@branch_owner_required  # Applying the custom decorator
 def team_create(request):
     if request.method == 'POST':
         form = TeamForm(request.POST)
@@ -16,11 +22,10 @@ def team_create(request):
             team.branch = request.user.branch
             team.owner = request.user
             team.save()
+            logger.info(f'Team {team.name} created by {request.user.username}')
             return redirect('team_list')
 
-
     form = TeamForm(request.POST or None)
-    print(form.errors)
     return render(request, 'org/team_form.html', {'form': form})
 
 
@@ -39,16 +44,25 @@ def team_update(request, pk):
         form = TeamForm(request.POST, instance=team)
         if form.is_valid():
             form.save()
+            logger.info(f'Team {team.name} updated by {request.user.username}')
             return redirect('team_list')
     else:
         form = TeamForm(instance=team)
     return render(request, 'team_form.html', {'form': form})
 
+
 # Delete a team
 @login_required
 def team_delete(request, pk):
     team = get_object_or_404(Team, pk=pk)
+
+    if team.members.count() > 1:  
+        messages.error(request, "You cannot delete the team as it still has members.")
+        return redirect('team_list')
+
     if request.method == 'POST':
         team.delete()
+        messages.success(request, 'Team deleted successfully.')
         return redirect('team_list')
+
     return render(request, 'team_confirm_delete.html', {'team': team})
